+++ ui-packages/packages/consoles-common/src/components/layout/AboutModalBox/__mocks__/AboutModalBox.tsx
