--- ui-packages/packages/task-console-webapp/codegen.yml
@@ -1,17 +1,0 @@
-overwrite: true
-schema: "http://localhost:8180/graphql"
-documents: './src/graphql/queries.tsx'
-generates:
-  ./src/graphql/types.tsx:
-    plugins:
-      - add: '/* tslint:disable */'
-      - "typescript"
-      - "typescript-operations"
-      - "typescript-react-apollo"
-    config:
-      withHOC: false
-      withHooks: true
-      withComponent: false
-  ./src/graphql/graphql.schema.json:
-    plugins:
-      - "introspection"
