+++ ui-packages/packages/task-console-webapp/src/components/pages/TaskDetailsPage/__mocks__/TaskDetailsPage.tsx
